requirements.txt
Click
Django
Markdown
asgiref
asn1crypto
certifi
cffi
chardet
cryptography
django-colorfield
django-crispy-forms
django-debug-toolbar
django-debug-toolbar-request-history
django-formtools
django-itemlist==0.2.10
django-memoize
django-model-utils
django-proxy
django-python3-ldap
geopy
habanero
idna
jsonfield2
ldap3
msgpack
numpy
pip-chill
psycopg2-binary
pyasn1
pycparser
python-dateutil
python-memcached
pytz
requests
setuptools
six
sqlparse
timezonefinder
urllib3
fastjsonschema